conditional compilation for cargo test
