+++ rust/threshold-schnorr/src/schnorr_example_rust/src/all_architectures.rs
@@ -1,0 +1,25 @@
+use candid::CandidType;
+use serde::{Deserialize, Serialize};
+
+#[derive(CandidType, Serialize, Deserialize, Debug, Copy, Clone)]
+pub enum SchnorrAlgorithm {
+    #[serde(rename = "bip340secp256k1")]
+    Bip340Secp256k1,
+    #[serde(rename = "ed25519")]
+    Ed25519,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct PublicKeyReply {
+    pub public_key_hex: String,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct SignatureReply {
+    pub signature_hex: String,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct SignatureVerificationReply {
+    pub is_signature_valid: bool,
+}
